drivers: ad9517: fix write operation for LVPECL
The address mapping of LVPECL output channel registers contains and
offset of 3 bytes between the first two channels and the last two
channels.

Even though the macro defines are correct, the current implementation of
the driver neglects this aspect.

This patch makes sure that the register addresses are correctly
composed when data is written.

Datasheet link (page 57 of 80):
https://www.analog.com/media/en/technical-documentation/data-sheets/AD9517-3.pdf

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>